Force poi check row 1 cell 0 and 1 if empty celltype or cellvalue skip row Change model validator check end with "-ES" instead of "ES"

+++ src/main/java/com/advantech/service/WorktimeService.java
@@ -180,6 +180,9 @@ public class WorktimeService {
 
     public int delete(int id) {
         Worktime worktime = this.findByPrimaryKey(id);
+        if (worktime == null) {
+            return 0;
+        }
         return worktimeDAO.delete(worktime);
     }
